(WIP) Add volatile_fact() function
This is a quick POC to add a function that resolves facts on the agent
side during catalog application. It's useful only when using cached
catalogs that depend on facts that might change over time.

This was accomplished by slightly modifying `fact()` so that it could be
deferred and then writing a quick Puppet code function to wrap it in a
`Deferred` object. If reviewers agree that this functionality is useful,
then I'll port it to a Ruby fact so the name doesn't have to be
namespaced and it can parallel the existing `fact()` function.

You would use it like so:

```
notify { 'testing':
  message  => stdlib::volatile_fact('some.volatile.fact'),
}
```