[MODULES-10729] defined_with_params - unnamed type
This fixes an issue where trying to match against any declared defined
type with a specific parameter would never return that there was a match
in the catalog.

This is relevant when trying to ensure that you don't have port
conflicts in services that can declare multiple instances of a service
through a given defined type, such as Apache.

This, previously valid call to the function now works as expected:

``defined_with_params(My::DefinedType, { 'port' => 1337 })``

MODULES-10729 #close