stdlib::ensure: new fuction to cast ensure values
This CR adds a new function to case the ensureable property to a resource
specific value. e.g.
  * stdlib::ensure('present', 'service') == 'running'
  * stdlib::ensure('present', 'directory') == 'directory'
This is usefull when you want to pass ensure to a custome class and
ensure all the resource get the correct value e.g.

```
class foo(
  Enum['present', 'absent'] $ensure = 'present'
) {
  file {'/some/dir':
    ensure => stdlib::ensure($ensure, 'directory')
  }
  file {'/some/file':
    ensure => stdlib::ensure($ensure, 'file')
  }
  file {'/some/link':
    ensure => stdlib::ensure($ensure, 'link')
  }
  service {'some-service':
    ensure => stdlib::ensure($ensure, 'service')
  }
```

something similar to this was discussed in #869