(MAINT) Limit GCG to generating changes since tag '3.0.0'
When attempting to run the GCG for a new release, we run in to
issues with the Github API as there is no common ancestor between
commit `b305bbeac7a0560a271f34026f936b88b88da477` and tag `2.1.3`.

Rather than try to fix up commits / tags from circa 2011/12, it
makes more sense to limit the GCG to 3.0.0 onwards.