Fix serialization of undef in to_python()
:undef was used in old Puppet functions, but modern Puppet 4 functions
receive `undef` Puppet values as `nil` Ruby values.  This cause
to_python() to serialize `undef` as `nil` instead of `None`, which is
not valid Python code.

Fix the to_python() behavior by comparing with `nil`.  Update the test
suite accordingly.

While here, also adjust the to_ruby() function which had the same wrong
logic but was not causing trouble because in Ruby, we want to serialize
`nil` to `nil` :-)