Rewrite validate_domain_name() as a Puppet 4.x function
The 3.x function rely on is_domain_name() which is deprecated. Rewrite
it using the more modern puppet 4.x function to rely on data types for
better parameters validation.

While here, adjust the Stdlib::Fqdn / Stdlib::Host data types to test
fully numerical addresses and add a Stdlib::Dns::Zonedata type that is
basically the same but with a trailing dot.