types - dns: add a new dns type with more specific fqdn definitions
This PR introduces a new set of types for validating FQDN's.  it creates
a type for the more loose rfc definitions of domain names and a stricter
and likely more useful iana type.

This allows us to create a new type that more does to what most users
expect i.e. that a dns name is one that works on the internt, without
breaking current uses for users that may be using the Stdlib::Fqdn to
validate validate DNS names that don't work with the IANA roots.

The intention of this patch would be to deprecate the currnet Stdlib::Fqdn
type and encourage users to move to the appropriate Stdlib::DNS::* type
which for most users will likely be the stricter Stdlib::DNS::Fqdn type

Note: this PR is intentionally a bit rough to first garner thoughts as to
if this is the correct direction

Fixes #1282 (not sure it fixes but want it tagged)