Fix validate_domain_name called without parameters
In #1345, the validate_domain_name() function was rewritten to the
Puppet 4.x syntax, but the behaviour was slightly changed since we
previously did not allow passing no parameter at all.

Add code to assert as least one paramatter is passed to restore the
previous behavior and raise an error when called without parameter.