Deprecate the `validate_legacy()` function
`validate_legacy()` was added to help migrating from the legacy
validation functions to the new Puppet data types which where not 100%
backward compatible with the previous validation (e.g. `'42'` was a
valid integer according to `validate_integer()` but indeed a String and
not an Integer when using data types.

The legacy validation functions have been removed from stdlib, so
`validate_legacy()` will now fail if it tries to run them.  But as
they produced deprecation warning, they are supposed to have already
been fixed.  For the sake of security, instead of removing
`validate_legacy()` now, deprecate it so that it is strictly equivalent
to validating using a data type, but also emits a warning.