Add `use_strict_setting` parameter to `deprecation` function
In this change, an optional third parameter is added to the
`deprecation` function.  The default behaviour remains the same, but
when the third parameter is set to `false`, Puppet's `strict` setting is
ignored. This means that even if `strict` is set to `error`, no
exception will be raised when this function is called, and a warning
will be logged instead.

This is a prerequisite for #1373