Deprecate `time` function
`time` was both an old API ruby function, and not needed since Puppet 4.

In this commit, the `parser` function is removed and replaced by a
non-namespaced APIv4 function. This is done to preserve compatibility
before the function is removed completely in a later release.

The ruby function calls a Puppet language function which implements the
existing behaviour using core Puppet functionality.

The original function did something weird with an optional `timezone`
parameter. Nobody really remembers why. It appears that very early
versions of Ruby may have been buggy and needed this as a workaround.

To be clear, seconds since the epoch are the same _everywhere_. Timezone
is not relevant.