Fix loading JSON from URL
Starting with Puppet 8, stdlib use the JSON class instead of the PSON
class that was used before to load JSON content downloaded from an URL.

The PSON class allowed to pass an IO object and #read its content for
parsing, but the JSON class mandates passing a String.

The loadjson() functions, when passed an URL, download the document with
OpenURI.open_uri (which return a StringIO object) and try to parse it.
On Puppet 7, this works because PSON accept IO objects, but with Puppet
8 an exception is raised:

```
Error: Evaluation Error: Error while evaluating a Function Call, no implicit conversion of StringIO into String
```

As the test suite inadequately mocked OpenURI.open_uri to return a
String, this was not catch.

Fix this by adjusting the return value in the test suite to match the
actual behavior of OpenURI and calling #read on the IO object returned
by OpenURI.open_uri.

While here, add an acceptance tests to check loadjson when using an URL.